% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penal.R
\name{penal}
\alias{penal}
\title{Computes matrices used for weighted group penalty}
\usage{
penal(n, t, lambda2)
}
\arguments{
\item{n}{size of sample data set}

\item{t}{survival times}

\item{lambda2}{smoothness parameter}
}
\value{
penalization matrices
}
\description{
This function takes the number of examples, a sequence of observed failure times t and penalization parameter
lambda2. Note that lambda2 controls the smoothness of the splines since we are regularizing
the second derivatives of B-spline basis functions. The output will be a t x t matrix of "weights"
}
