% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.prx.R
\name{plot.prx}
\alias{plot.prx}
\title{FUNCTION_TITLE}
\usage{
\method{plot}{prx}(
  obj,
  alpha = 0.05,
  xlab = "time",
  ylab = "X effect",
  all.terms = TRUE,
  variable,
  continue = FALSE,
  col,
  b.theta = TRUE
)
}
\arguments{
\item{obj}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05}

\item{xlab}{PARAM_DESCRIPTION, Default: 'time'}

\item{ylab}{PARAM_DESCRIPTION, Default: 'X effect'}

\item{all.terms}{PARAM_DESCRIPTION, Default: TRUE}

\item{variable}{PARAM_DESCRIPTION}

\item{continue}{PARAM_DESCRIPTION, Default: FALSE}

\item{col}{PARAM_DESCRIPTION}

\item{b.theta}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
