% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{prep_data}
\usage{
prep_data(time, cens, X, num_knots, center = FALSE)
}
\arguments{
\item{time}{survival time (either inorder or unordered)}

\item{cens}{censored indicator (binary vector)}

\item{X}{covariates of interest}

\item{num_knots}{number of knots to use in spline}

\item{center}{centering, Default: FALSE}
}
\value{
list containg processed data (time, cens, X) and basis functions evaluated at each time point, knots used
}
\description{
outputs list of processes survival data necissary for prox.fit and reg.fit
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[mgcv]{smooth.construct.cr.smooth.spec}}
}
