% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.fit.R
\name{prox.fit}
\alias{prox.fit}
\title{prox.fit}
\usage{
prox.fit(
  time,
  death,
  X,
  names = NULL,
  Ftime,
  knots,
  theta,
  iter.max,
  lambda1,
  lambda2,
  ttime,
  acceleration = FALSE,
  delta = 0.5,
  verbal = FALSE
)
}
\arguments{
\item{time}{PARAM_DESCRIPTION}

\item{death}{PARAM_DESCRIPTION}

\item{X}{PARAM_DESCRIPTION}

\item{names}{PARAM_DESCRIPTION, Default: NULL}

\item{Ftime}{PARAM_DESCRIPTION}

\item{knots}{PARAM_DESCRIPTION}

\item{theta}{PARAM_DESCRIPTION}

\item{iter.max}{PARAM_DESCRIPTION}

\item{lambda1}{PARAM_DESCRIPTION}

\item{lambda2}{PARAM_DESCRIPTION}

\item{ttime}{PARAM_DESCRIPTION}

\item{acceleration}{PARAM_DESCRIPTION, Default: FALSE}

\item{delta}{PARAM_DESCRIPTION, Default: 0.5}

\item{verbal}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
