% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxCox-package.R
\docType{package}
\name{ProxCox-package}
\alias{ProxCox}
\alias{ProxCox-package}
\title{ProxCox: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jacob Shkrob \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
