% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumevents_prx.R
\name{sumevents_prx}
\alias{sumevents_prx}
\title{helper function for Partial Cox Likelihood calculation for proximal method}
\usage{
sumevents_prx(i, X, Ftime, theta, n, p, eventtimes)
}
\arguments{
\item{i}{current index}

\item{X}{feature matrix}

\item{Ftime}{matrix of values of the basis for certain time points}

\item{theta}{current parameters of partial Cox likelihood}

\item{n}{dimension}

\item{p}{number of variables}

\item{eventtimes}{indexes of individuals that had event of interest}
}
\description{
helper function for Partial Cox Likelihood calculation for proximal method
}
