% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{FUNCTION_TITLE}
\usage{
prep_data(time, cens, X, num_knots, center = FALSE)
}
\arguments{
\item{time}{PARAM_DESCRIPTION}

\item{cens}{PARAM_DESCRIPTION}

\item{X}{PARAM_DESCRIPTION}

\item{num_knots}{PARAM_DESCRIPTION}

\item{center}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[mgcv]{smooth.construct.cr.smooth.spec}}
}
