% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.fit.R
\name{prox.fit}
\alias{prox.fit}
\title{prox.fit}
\usage{
prox.fit(
  time,
  death,
  X,
  names = NULL,
  Ftime,
  knots,
  theta,
  iter.max,
  lambda1,
  lambda2,
  ttime,
  acceleration = FALSE,
  delta = 0.5,
  verbal = FALSE
)
}
\arguments{
\item{time}{time-to-event data from prep_data}

\item{death}{survival information; binary vector from prep_data}

\item{X}{covariates of interest from prep_data}

\item{names}{column names of the covariates X, Default: NULL}

\item{Ftime}{B-spline basis functions evaluated at each survival time point from prep_data}

\item{knots}{knots of B-spline basis from prep_data}

\item{theta}{optional initialization of theta}

\item{iter.max}{iteration number for proximal gradient method}

\item{lambda1}{regularization parameter for sparsity}

\item{lambda2}{regularization parameter for smoothness penalty}

\item{ttime}{initial learning rate}

\item{acceleration}{use Nesterov acceleration, Default: FALSE}

\item{delta}{backtracking step size, Default: 0.5}

\item{verbal}{output of information at each iteration, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Outputs B-spline basis coefficients for each covariate of interest using Proximal Gradient Descent method on penalized Cox Partial Likelihood function
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
